# linkedin_connector/docs/uat_evidence/job_orchestrator_p1_p3_uat_20260803T044436Z/worker_openapi.json
{"openapi":"3.1.0","info":{"title":"Personal Job Apply Worker","description":"Dry-run Playwright worker for offline ATS fixtures. Real submit is hard-disabled. LinkedIn URLs are rejected.","version":"0.1.0"},"paths":{"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthResponse"}}}}}}},"/v1/apply/draft":{"post":{"summary":"Apply Draft","operationId":"apply_draft_v1_apply_draft_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApplyDraftRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApplyAttemptResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/apply/submit":{"post":{"summary":"Apply Submit","description":"Hard-disabled in this phase — always 403.","operationId":"apply_submit_v1_apply_submit_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApplySubmitRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/apply/{attempt_id}":{"get":{"summary":"Get Attempt","operationId":"get_attempt_v1_apply__attempt_id__get","parameters":[{"name":"attempt_id","in":"path","required":true,"schema":{"type":"string","title":"Attempt Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApplyAttemptResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"ApplicantFixture":{"properties":{"full_name":{"type":"string","maxLength":120,"title":"Full Name","default":"Alex Example"},"email":{"type":"string","maxLength":200,"title":"Email","default":"alex.example@example.test"},"phone":{"type":"string","maxLength":40,"title":"Phone","default":"+10000000000"},"cover_letter":{"type":"string","maxLength":5000,"title":"Cover Letter","default":"This is a fictional dry-run cover letter for offline fixture testing."},"cv_filename":{"type":"string","pattern":"^test\\.pdf$","title":"Cv Filename","default":"test.pdf"}},"type":"object","title":"ApplicantFixture","description":"Fictional applicant data only — never real credentials or CV paths."},"ApplyAttemptResponse":{"properties":{"attempt_id":{"type":"string","title":"Attempt Id"},"state":{"$ref":"#/components/schemas/ApplyState"},"stop_reason":{"$ref":"#/components/schemas/StopReason"},"dry_run":{"type":"boolean","title":"Dry Run","default":true},"final_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Final Url"},"screenshot_paths":{"items":{"type":"string"},"type":"array","title":"Screenshot Paths"},"adapter_used":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Adapter Used"},"filled_fields":{"items":{"type":"string"},"type":"array","title":"Filled Fields"},"message":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Message"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"},"metadata":{"additionalProperties":true,"type":"object","title":"Metadata"}},"type":"object","required":["attempt_id","state","stop_reason","created_at","updated_at"],"title":"ApplyAttemptResponse"},"ApplyDraftRequest":{"properties":{"url":{"type":"string","title":"Url","description":"Target URL. Must be file:// fixture or offline HTML path."},"dry_run":{"type":"boolean","title":"Dry Run","description":"Must be true. Real submits are hard-disabled."},"applicant":{"$ref":"#/components/schemas/ApplicantFixture"},"adapter":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Adapter","description":"Optional adapter hint: bebee_like | greenhouse_like | auto"},"metadata":{"additionalProperties":true,"type":"object","title":"Metadata"}},"type":"object","required":["url","dry_run"],"title":"ApplyDraftRequest","description":"POST /v1/apply/draft body."},"ApplyState":{"type":"string","enum":["pending","running","drafted","stopped","failed","submit_disabled"],"title":"ApplyState"},"ApplySubmitRequest":{"properties":{"attempt_id":{"type":"string","title":"Attempt Id"},"dry_run":{"type":"boolean","title":"Dry Run","default":true},"confirm":{"type":"boolean","title":"Confirm","default":false}},"type":"object","required":["attempt_id"],"title":"ApplySubmitRequest","description":"POST /v1/apply/submit body — always rejected with 403."},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HealthResponse":{"properties":{"status":{"type":"string","title":"Status","default":"ok"},"service":{"type":"string","title":"Service","default":"personal-job-apply-worker"},"dry_run_only":{"type":"boolean","title":"Dry Run Only","default":true},"submit_enabled":{"type":"boolean","title":"Submit Enabled","default":false},"version":{"type":"string","title":"Version","default":"0.1.0"}},"type":"object","title":"HealthResponse"},"StopReason":{"type":"string","enum":["captcha","otp","login_wall","consent","unknown_question","linkedin_blocked","submit_disabled","dry_run_required","invalid_url","none"],"title":"StopReason"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}